function/requirements.txt
azure-functions
sendgrid
psycopg2-binary==2.8.6